+++ src/main/java/com/gymapp/gymapp/Repository/ExerciseRepository.java
@@ -3,5 +3,5 @@ package com.gymapp.gymapp.Repository;
 import com.gymapp.gymapp.Entities.Exercise;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExerciseRepository extends JpaRepository<Exercise, Integer> {
+public interface ExerciseRepository extends JpaRepository<Exercise, Long> {
 }
